feat: added the StaffNote component and refactored cube cards
