+++ src/lib/components/cube/CubeCardSkeleton.svelte
